Rename `query` module
